+++ service/controller/cluster_resource_set.go
@@ -22,7 +22,6 @@ import (
 	"github.com/giantswarm/aws-operator/pkg/project"
 	"github.com/giantswarm/aws-operator/service/controller/controllercontext"
 	"github.com/giantswarm/aws-operator/service/controller/internal/changedetection"
-	"github.com/giantswarm/aws-operator/service/controller/internal/cloudconfig"
 	"github.com/giantswarm/aws-operator/service/controller/internal/encrypter"
 	"github.com/giantswarm/aws-operator/service/controller/internal/encrypter/kms"
 	"github.com/giantswarm/aws-operator/service/controller/key"
@@ -37,6 +36,7 @@ import (
 	"github.com/giantswarm/aws-operator/service/controller/resource/cleanupsecuritygroups"
 	"github.com/giantswarm/aws-operator/service/controller/resource/cproutetables"
 	"github.com/giantswarm/aws-operator/service/controller/resource/cpvpc"
+	"github.com/giantswarm/aws-operator/service/controller/resource/encryptionensurer"
 	"github.com/giantswarm/aws-operator/service/controller/resource/endpoints"
 	"github.com/giantswarm/aws-operator/service/controller/resource/eniconfigcrs"
 	"github.com/giantswarm/aws-operator/service/controller/resource/ensurecpcrs"
@@ -45,12 +45,11 @@ import (
 	"github.com/giantswarm/aws-operator/service/controller/resource/peerrolearn"
 	"github.com/giantswarm/aws-operator/service/controller/resource/region"
 	"github.com/giantswarm/aws-operator/service/controller/resource/s3bucket"
-	"github.com/giantswarm/aws-operator/service/controller/resource/s3object"
 	"github.com/giantswarm/aws-operator/service/controller/resource/secretfinalizer"
 	"github.com/giantswarm/aws-operator/service/controller/resource/service"
+	"github.com/giantswarm/aws-operator/service/controller/resource/snapshotid"
 	"github.com/giantswarm/aws-operator/service/controller/resource/tccp"
 	"github.com/giantswarm/aws-operator/service/controller/resource/tccpazs"
-	"github.com/giantswarm/aws-operator/service/controller/resource/tccpencryption"
 	"github.com/giantswarm/aws-operator/service/controller/resource/tccpf"
 	"github.com/giantswarm/aws-operator/service/controller/resource/tccpi"
 	"github.com/giantswarm/aws-operator/service/controller/resource/tccpoutputs"
@@ -111,35 +110,6 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 		}
 
 		encrypterObject, err = kms.NewEncrypter(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	var tccpCloudConfig *cloudconfig.TCCP
-	{
-		c := cloudconfig.TCCPConfig{
-			Config: cloudconfig.Config{
-				Encrypter: encrypterObject,
-				Logger:    config.Logger,
-
-				CalicoCIDR:                config.CalicoCIDR,
-				CalicoMTU:                 config.CalicoMTU,
-				CalicoSubnet:              config.CalicoSubnet,
-				ClusterIPRange:            config.ClusterIPRange,
-				DockerDaemonCIDR:          config.DockerDaemonCIDR,
-				IgnitionPath:              config.IgnitionPath,
-				ImagePullProgressDeadline: config.ImagePullProgressDeadline,
-				ClusterDomain:             config.ClusterDomain,
-				NetworkSetupDockerImage:   config.NetworkSetupDockerImage,
-				PodInfraContainerImage:    config.PodInfraContainerImage,
-				RegistryDomain:            config.RegistryDomain,
-				SSHUserList:               config.SSHUserList,
-				SSOPublicKey:              config.SSOPublicKey,
-			},
-		}
-
-		tccpCloudConfig, err = cloudconfig.NewTCCP(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -258,6 +228,18 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 		}
 	}
 
+	var snapshotIDResource resource.Interface
+	{
+		c := snapshotid.Config{
+			Logger: config.Logger,
+		}
+
+		snapshotIDResource, err = snapshotid.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	var tccpAZsResource resource.Interface
 	{
 		c := tccpazs.Config{
@@ -274,14 +256,14 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 		}
 	}
 
-	var tccpEncryptionResource resource.Interface
+	var encryptionEnsurerResource resource.Interface
 	{
-		c := tccpencryption.Config{
+		c := encryptionensurer.Config{
 			Encrypter: encrypterObject,
 			Logger:    config.Logger,
 		}
 
-		tccpEncryptionResource, err = tccpencryption.New(c)
+		encryptionEnsurerResource, err = encryptionensurer.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -354,30 +336,6 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 		}
 
 		s3BucketResource, err = toCRUDResource(config.Logger, ops)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	var s3ObjectResource resource.Interface
-	{
-		c := s3object.Config{
-			CertsSearcher:      config.CertsSearcher,
-			CloudConfig:        tccpCloudConfig,
-			LabelsFunc:         key.KubeletLabelsTCCP,
-			Logger:             config.Logger,
-			G8sClient:          config.K8sClient.G8sClient(),
-			PathFunc:           key.S3ObjectPathTCCP,
-			RandomKeysSearcher: config.RandomKeysSearcher,
-			RegistryDomain:     config.RegistryDomain,
-		}
-
-		ops, err := s3object.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-
-		s3ObjectResource, err = toCRUDResource(config.Logger, ops)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -486,6 +444,7 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 			Logger: config.Logger,
 
 			Route53Enabled: config.Route53Enabled,
+			ToClusterFunc:  key.ToCluster,
 		}
 
 		tccpOutputsResource, err = tccpoutputs.New(c)
@@ -706,6 +665,7 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 		// All these resources only fetch information from remote APIs and put them
 		// into the controller context.
 		awsClientResource,
+		snapshotIDResource,
 		accountIDResource,
 		natGatewayAddressesResource,
 		peerRoleARNResource,
@@ -719,13 +679,12 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 
 		// All these resources implement certain business logic and operate based on
 		// the information given in the controller context.
+		encryptionEnsurerResource,
 		apiEndpointResource,
 		ipamResource,
 		bridgeZoneResource,
-		tccpEncryptionResource,
 		tccpSecurityGroupsResource,
 		s3BucketResource,
-		s3ObjectResource,
 		tccpAZsResource,
 		tccpiResource,
 		tccpResource,