+++ service/controller/internal/cloudconfig/tccp.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 
 	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/pkg/apis/infrastructure/v1alpha2"
-	g8sv1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
 	"github.com/giantswarm/certs"
 	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v6/v_6_0_0"
 	"github.com/giantswarm/microerror"
@@ -87,12 +86,11 @@ func (t *TCCP) Render(ctx context.Context, cr infrastructurev1alpha2.AWSCluster,
 		params.DisableIngressControllerService = true
 		params.EtcdPort = key.EtcdPort
 		params.Extension = &MasterExtension{
-			awsConfigSpec: cmaClusterToG8sConfig(t.config, cr, labels),
 			baseExtension: baseExtension{
-				registryDomain: t.config.RegistryDomain,
 				cluster:        cr,
 				encrypter:      t.config.Encrypter,
 				encryptionKey:  cc.Status.TenantCluster.Encryption.Key,
+				registryDomain: t.config.RegistryDomain,
 			},
 			cc:               cc,
 			clusterCerts:     clusterCerts,
@@ -136,7 +134,6 @@ func (t *TCCP) Render(ctx context.Context, cr infrastructurev1alpha2.AWSCluster,
 }
 
 type MasterExtension struct {
-	awsConfigSpec g8sv1alpha1.AWSConfigSpec
 	baseExtension
 	// TODO Pass context to k8scloudconfig rendering fucntions
 	//
@@ -311,7 +308,7 @@ func (e *MasterExtension) Files() ([]k8scloudconfig.FileAsset, error) {
 
 	var fileAssets []k8scloudconfig.FileAsset
 
-	data := e.templateData()
+	data := e.templateDataTCCP()
 
 	for _, fm := range filesMeta {
 		c, err := k8scloudconfig.RenderFileAssetContent(fm.AssetContent, data)
@@ -393,8 +390,10 @@ func (e *MasterExtension) Units() ([]k8scloudconfig.UnitAsset, error) {
 
 	var newUnits []k8scloudconfig.UnitAsset
 
+	data := e.templateDataTCCP()
+
 	for _, fm := range unitsMeta {
-		c, err := k8scloudconfig.RenderAssetContent(fm.AssetContent, e.awsConfigSpec)
+		c, err := k8scloudconfig.RenderAssetContent(fm.AssetContent, data)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}