+++ service/controller/internal/cloudconfig/types.go
@@ -1,7 +1,13 @@
 package cloudconfig
 
-type templateData struct {
-	AWSRegion      string
-	IsChinaRegion  bool
-	RegistryDomain string
+type TemplateData struct {
+	AWSRegion            string
+	IsChinaRegion        bool
+	MasterENIAddress     string
+	MasterENIGateway     string
+	MasterENIName        string
+	MasterENISubnetSize  int
+	MasterEtcdVolumeName string
+	MasterID             int
+	RegistryDomain       string
 }