+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add mixed instance support for worker ASGs.
+
+### Changed
+
+- Improve cleanup of `DrainerConfig` CRs after node draining.
+- Use release.Revision in Helm chart for Helm 3 support.
+
 
 
 ## [8.3.0] 2020-04-17